requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.27
django_celery_results==2.6.0
kombu==5.6.2
numpy==2.4.1
packaging==26.0
pandas==3.0.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.3.1
